+++ R/utilities-data-combined.R
@@ -1,81 +1,49 @@
-#' Count number of objects
-#'
-#' @details
-#'
-#' If the argument is not a vector, unlike `length()`, this function will not
-#' count the number of named bindings in an environment, but only the number of
-#' instances of a class.
-#'
-#' For example, `length(mtcars)` will return 11, but `objCount(mtcars)` will
-#' return 1.
-#'
-#' @param x An object (an atomic vector, a list, or instance(s) of a class).
-#'
-#' @examples
-#'
-#' objCount(c(1, 2, 3)) # 3
-#' objCount(list("a", "b")) # 2
-#' objCount(mtcars) # 1
-#' @return Integer representing the count of objects.
-#'
-#' @keywords internal
-#' @noRd
-objCount <- function(x) {
-  # `is.vector()` can handle both atomic vectors and lists, i.e.
-  # both `is.vector(c(1, 2))` and `is.vector(list(1, 2))` will be `TRUE`
-  #
-  # For anything other than a vector, the object count should be 1
-  if (!is.vector(x)) {
-    return(1L)
-  }
-
-  return(length(x))
-}
-
 #'  Validate arguments provided as vectors
 #'
 #' @details
 #'
-#' Validation of arguments provided as a vector involves:
+#' Cleaning an argument provided as (atomic or generic) vector involves:
 #'
 #' - Checking that it is of expected length.
-#' - Checking for `NULL` or other unexpected values (`NaN`, `Inf`, `NA` of the
-#'   wrong type) and standardizing them to `NA` of desired type.
-#' - Checking that each element in the vector is of expected type.
-#' - If a non-atomic list is provided, converting it to an atomic vector.
+#' - Checking for `NULL` or other special constants (`NaN`, `Inf`, `NA` of the
+#'   wrong type) and standardizing them to `NA` of desired data type.
+#' - Checking that each element in the vector is of expected data type.
+#' - Making sure that an atomic vector is always returned, irrespective of if
+#'   the input was a list or an atomic vector.
 #'
 #' @param x A vector of arguments.
 #' @param expectedLength An integer to denote the expected length of the vector.
-#' @inheritParams flattenList
+#' @inheritParams ospsuite.utils::validateIsOfType
 #'
 #' @return
 #'
-#' An atomic vector of desired type containing specified arguments.
+#' An atomic vector of desired data type.
 #'
 #' @examples
 #'
-#' cleanVectorArgs(list(1, 2, NA, NULL), 4L, "numeric")
-#' cleanVectorArgs(c(1, 2, NA, NA_complex), 4L, "numeric")
+#' ospsuite:::.cleanVectorArgs(list(1, 2, NA, NULL), 4L, "numeric")
+#' ospsuite:::.cleanVectorArgs(c(1, 2, NA, NA_complex), 4L, "numeric")
+#'
 #' @keywords internal
 #' @noRd
-cleanVectorArgs <- function(arg = NULL, expectedLength = NULL, type) {
-  # return early if argument was not specified
+.cleanVectorArgs <- function(arg = NULL, expectedLength = NULL, type) {
+  # Return early if no argument was specified
   if (is.null(arg)) {
     return(NULL)
   }
+
+  # Check that the argument is not empty
+  validateIsNotEmpty(arg)
 
   # validate the length of vector arguments
   if (!is.null(expectedLength)) {
     validateIsOfLength(arg, expectedLength)
   }
 
-  # validate depth of the vector
-  validateVecDepth(arg)
-
   # convert `NULL`s or logical `NA`s to `NA` of required type
 
   # Note that `purrr::map()` will return a list
-  arg <- purrr::map(arg, ~ toMissingOfType(.x, type))
+  arg <- purrr::map(arg, ~ .toMissingOfType(.x, type))
 
   # validate the type of arguments
 
@@ -85,7 +53,7 @@ cleanVectorArgs <- function(arg = NULL, expectedLength = NULL, type) {
 
   # arguments are still in a list
   # flatten them to an atomic vector of required type
-  arg <- flattenList(arg, type)
+  arg <- .flattenList(arg, type)
 
   return(arg)
 }
@@ -102,15 +70,16 @@ cleanVectorArgs <- function(arg = NULL, expectedLength = NULL, type) {
 #'
 #' @examples
 #'
-#' flattenList(list(1, 2, 3, NA), type = "numeric")
-#' flattenList(list(TRUE, FALSE, NA), type = "integer")
+#' ospsuite:::.flattenList(list(1, 2, 3, NA), type = "numeric")
+#' ospsuite:::.flattenList(list(TRUE, FALSE, NA), type = "integer")
+#'
 #' @return An atomic vector of desired type.
 #'
 #' @keywords internal
 #' @noRd
-flattenList <- function(x, type) {
-  if (!is.vector(x)) {
-    stop("`x` argument can only be a vector.", call. = FALSE)
+.flattenList <- function(x, type) {
+  if (!is.null(dim(x))) {
+    stop("Argument to parameter `x` can only be a vector.")
   }
 
   if (is.list(x)) {
@@ -132,15 +101,16 @@ flattenList <- function(x, type) {
 #' Convert `NULL` or `NA`s to `NA` of desired type
 #'
 #' @param x A single element.
-#' @inheritParams flattenList
+#' @inheritParams .flattenList
 #'
 #' @examples
 #'
-#' toMissingOfType(NA, type = "real")
-#' toMissingOfType(NULL, type = "integer")
+#' ospsuite:::.toMissingOfType(NA, type = "real")
+#' ospsuite:::.toMissingOfType(NULL, type = "integer")
+#'
 #' @keywords internal
 #' @noRd
-toMissingOfType <- function(x, type) {
+.toMissingOfType <- function(x, type) {
   # all unexpected values will be converted to `NA` of a desired type
   if (is.null(x) || is.na(x) || is.nan(x) || is.infinite(x)) {
     x <- switch(type,
@@ -156,39 +126,4 @@ toMissingOfType <- function(x, type) {
   }
 
   return(x)
-}
-
-#' Check if the vector depth is as expected
-#'
-#' @param x A vector whose depth needs to be checked.
-#'
-#' @description
-#'
-#' For function arguments that accept a vector, a vector with depth greater than
-#' 2 is rarely acceptable. This function will produce an error if this is the
-#' case.
-#'
-#' Vector depths:
-#
-# 1 = atomic vector (or empty list)
-# 2 = non-nested list
-# > 2 = nested list
-#'
-#' @examples
-#'
-#' validateVecDepth(c(1)) # depth is 1
-#' validateVecDepth(list()) # depth is 1
-#' validateVecDepth(list(1)) # depth is 2
-#'
-#' # this will produce an error
-#' # validateVecDepth(list(list(1))) # depth is 3
-#' @keywords internal
-#' @noRd
-validateVecDepth <- function(x) {
-  if (purrr::vec_depth(x) > 2L) {
-    stop(
-      "A nested list is not a valid argument here.",
-      call. = FALSE
-    )
-  }
 }