+++ R/utilities-data-set.R
@@ -73,31 +73,27 @@ saveDataSetToPKML <- function(dataSet, filePath) {
 #' `DataSet`.
 #'
 #' @export
-
 dataSetToDataFrame <- function(dataSets) {
   dataSets <- c(dataSets)
   validateIsOfType(dataSets, "DataSet")
 
-  # styler: off
-  name         <- .makeDataFrameColumn(dataSets, "name")
-  xUnit        <- .makeDataFrameColumn(dataSets, "xUnit")
-  yUnit        <- .makeDataFrameColumn(dataSets, "yUnit")
-  yErrorUnit   <- .makeDataFrameColumn(dataSets, "yErrorUnit")
-  xDimension   <- .makeDataFrameColumn(dataSets, "xDimension")
-  yDimension   <- .makeDataFrameColumn(dataSets, "yDimension")
-  yErrorType   <- .makeDataFrameColumn(dataSets, "yErrorType")
-  molWeight    <- .makeDataFrameColumn(dataSets, "molWeight")
-  xValues      <- .makeDataFrameColumn(dataSets, "xValues")
-  yValues      <- .makeDataFrameColumn(dataSets, "yValues")
+  name <- .makeDataFrameColumn(dataSets, "name")
+  xUnit <- .makeDataFrameColumn(dataSets, "xUnit")
+  yUnit <- .makeDataFrameColumn(dataSets, "yUnit")
+  yErrorUnit <- .makeDataFrameColumn(dataSets, "yErrorUnit")
+  xDimension <- .makeDataFrameColumn(dataSets, "xDimension")
+  yDimension <- .makeDataFrameColumn(dataSets, "yDimension")
+  yErrorType <- .makeDataFrameColumn(dataSets, "yErrorType")
+  molWeight <- .makeDataFrameColumn(dataSets, "molWeight")
+  xValues <- .makeDataFrameColumn(dataSets, "xValues")
+  yValues <- .makeDataFrameColumn(dataSets, "yValues")
   yErrorValues <- .makeDataFrameColumn(dataSets, "yErrorValues")
-  lloq         <- .makeDataFrameColumn(dataSets, "LLOQ")
-  # styler: on
+  lloq <- .makeDataFrameColumn(dataSets, "LLOQ")
 
   df <- data.frame(
     name, xValues, yValues, yErrorValues, xDimension, xUnit, yDimension,
     yUnit, yErrorType, yErrorUnit, molWeight, lloq
   )
-
 
   # get all names of meta data entries from all data sets
   metaDataNames <- unique(unlist(lapply(dataSets, function(dataSets) {
@@ -110,16 +106,27 @@ dataSetToDataFrame <- function(dataSets) {
     df[[name]] <- col
   }
 
+  # consistently return a (classical) data frame
   return(df)
 }
 
+#' @rdname dataSetToDataFrame
+#'
+#' @export
+dataSetToTibble <- function(dataSets) {
+  df <- dataSetToDataFrame(dataSets)
+
+  # consistently return a tibble data frame
+  return(dplyr::as_tibble(df))
+}
 
 #' Load data sets from excel
 #'
 #' @details Load observed data from an excel file using an importer configuration
 #'
 #' @param xlsFilePath Path to the excel file with the data
-#' @param importerConfiguration An object of type `DataImporterConfiguration` that is valid for the excel file
+#' @param importerConfigurationOrPath An object of type `DataImporterConfiguration` that is valid
+#'  for the excel file or a path to a XML file with stored configuration
 #' @param importAllSheets If `FALSE` (default), only sheets specified in the
 #' `importerConfiguration` will be loaded. If `TRUE`, an attempt to load all sheets
 #' is performed. If any sheet does not comply with the configuration, an error is thrown.
@@ -135,11 +142,23 @@ dataSetToDataFrame <- function(dataSets) {
 #'
 #' dataSets <- loadDataSetsFromExcel(
 #'   xlsFilePath = xlsFilePath,
-#'   importerConfiguration = importerConfiguration,
+#'   importerConfigurationOrPath = importerConfiguration,
 #'   importAllSheets = FALSE
 #' )
-loadDataSetsFromExcel <- function(xlsFilePath, importerConfiguration, importAllSheets = FALSE) {
+#'
+#' importerConfigurationFilePath <- system.file("extdata", "dataImporterConfiguration.xml", package = "ospsuite")
+#'
+#' dataSets <- loadDataSetsFromExcel(
+#'   xlsFilePath = xlsFilePath,
+#'   importerConfigurationOrPath = importerConfigurationFilePath,
+#'   importAllSheets = FALSE
+#' )
+loadDataSetsFromExcel <- function(xlsFilePath, importerConfigurationOrPath, importAllSheets = FALSE) {
   validateIsString(xlsFilePath)
+  importerConfiguration <- importerConfigurationOrPath
+  if (is.character(importerConfigurationOrPath)) {
+    importerConfiguration <- loadDataImporterConfiguration(importerConfigurationOrPath)
+  }
   validateIsOfType(importerConfiguration, "DataImporterConfiguration")
   validateIsLogical(importAllSheets)